+++ src/Api/WebSocket.hs
@@ -4,7 +4,10 @@ import Control.Monad (forever)
 import Control.Monad.IO.Class (liftIO)
 import Control.Monad.Logger (logErrorNS)
 import Control.Monad.Reader.Has (ask)
+import Data.Aeson ((.=))
 import Data.Aeson qualified as Aeson
+import Data.Bifunctor (bimap)
+import Data.Map qualified as Map
 import Data.Text (Text)
 import Data.Text qualified as Text
 import Network.WebSockets qualified as WebSockets
@@ -47,6 +50,7 @@ import Database (
   DatabaseError (DatabaseErrorDecodeError, DatabaseErrorNotFound),
  )
 import Database qualified
+import PlutusData (Data)
 
 type WSResponse =
   Either
@@ -75,11 +79,17 @@ getDatumsByHashes ::
   App WSResponse
 getDatumsByHashes hashes = do
   res <- Database.getDatumsByHashes hashes
-  pure $ case res of
-    Left err ->
-      Left $ mkGetDatumsByHashesFault $ Text.pack $ show err
-    Right datums ->
-      Right $ mkGetDatumsByHashesResponse (Just $ Aeson.toJSON <$> [datums])
+  pure $
+    bimap
+      (mkGetDatumsByHashesFault . Text.pack . show)
+      (mkGetDatumsByHashesResponse . encodeMap)
+      res
+  where
+    encodeMap :: Map.Map DataHash (Either DatabaseError Data) -> [Aeson.Value]
+    encodeMap datums = encodeOne <$> Map.toList datums
+
+    encodeOne :: (DataHash, Either DatabaseError Data) -> Aeson.Value
+    encodeOne (hash, datum) = Aeson.object ["hash" .= hash, "value" .= datum]
 
 getTxByHash ::
   Text ->